# tlf_journal: put block data to journal outside of lock