--- a/go/kbfs/libkbfs/block_journal.go
+++ b/go/kbfs/libkbfs/block_journal.go
@@ -346,22 +346,24 @@ func (j *blockJournal) end() (journalOrdinal, error) {
 	return last + 1, nil
 }
 
-func (j *blockJournal) hasData(id kbfsblock.ID) (bool, error) {
-	return j.s.hasData(id)
+func (j *blockJournal) hasData(
+	ctx context.Context, id kbfsblock.ID) (bool, error) {
+	return j.s.hasData(ctx, id)
 }
 
-func (j *blockJournal) isUnflushed(id kbfsblock.ID) (bool, error) {
-	return j.s.isUnflushed(id)
+func (j *blockJournal) isUnflushed(
+	ctx context.Context, id kbfsblock.ID) (bool, error) {
+	return j.s.isUnflushed(ctx, id)
 }
 
 func (j *blockJournal) remove(ctx context.Context, id kbfsblock.ID) (
 	removedBytes, removedFiles int64, err error) {
-	bytesToRemove, err := j.s.getDataSize(id)
+	bytesToRemove, err := j.s.getDataSize(ctx, id)
 	if err != nil {
 		return 0, 0, err
 	}
 
-	err = j.s.remove(id)
+	err = j.s.remove(ctx, id)
 	if err != nil {
 		return 0, 0, err
 	}
@@ -380,18 +382,20 @@ func (j *blockJournal) empty() bool {
 	return j.j.empty() && j.deferredGC.empty()
 }
 
-func (j *blockJournal) getDataWithContext(id kbfsblock.ID, context kbfsblock.Context) (
+func (j *blockJournal) getDataWithContext(
+	ctx context.Context, id kbfsblock.ID, context kbfsblock.Context) (
 	[]byte, kbfscrypto.BlockCryptKeyServerHalf, error) {
-	return j.s.getDataWithContext(id, context)
+	return j.s.getDataWithContext(ctx, id, context)
 }
 
-func (j *blockJournal) getData(id kbfsblock.ID) (
+func (j *blockJournal) getData(ctx context.Context, id kbfsblock.ID) (
 	[]byte, kbfscrypto.BlockCryptKeyServerHalf, error) {
-	return j.s.getData(id)
+	return j.s.getData(ctx, id)
 }
 
-func (j *blockJournal) getDataSize(id kbfsblock.ID) (int64, error) {
-	return j.s.getDataSize(id)
+func (j *blockJournal) getDataSize(
+	ctx context.Context, id kbfsblock.ID) (int64, error) {
+	return j.s.getDataSize(ctx, id)
 }
 
 func (j *blockJournal) getStoredBytes() int64 {
@@ -406,9 +410,9 @@ func (j *blockJournal) getStoredFiles() int64 {
 	return j.aggregateInfo.StoredFiles
 }
 
-// putData puts the given block data. If err is non-nil, putData will
+// putBlockData puts the given block data. If err is non-nil, putData will
 // always be false.
-func (j *blockJournal) putData(
+func (j *blockJournal) putBlockData(
 	ctx context.Context, id kbfsblock.ID, context kbfsblock.Context,
 	buf []byte, serverHalf kbfscrypto.BlockCryptKeyServerHalf) (
 	putData bool, err error) {
@@ -422,33 +426,44 @@ func (j *blockJournal) putData(
 		}
 	}()
 
+	putData, err = j.s.put(ctx, true, id, context, buf, serverHalf, "")
+	if err != nil {
+		return false, err
+	}
+
+	return putData, nil
+}
+
+// appendBlock appends an entry for the previously-put block to the
+// journal, and records the size for the put block.
+func (j *blockJournal) appendBlock(
+	ctx context.Context, id kbfsblock.ID, context kbfsblock.Context,
+	bufLenToAdd int64) error {
+	j.log.CDebugf(ctx, "Appending %block %s to journal", id)
+
+	if bufLenToAdd > 0 {
+		var putFiles int64 = filesPerBlockMax
+		err := j.accumulateBlock(bufLenToAdd, putFiles)
+		if err != nil {
+			return err
+		}
+	}
+
 	next, err := j.next()
 	if err != nil {
-		return false, err
+		return err
 	}
 
-	putData, err = j.s.put(true, id, context, buf, serverHalf, next.String())
+	err = j.s.addReference(ctx, id, context, next.String())
 	if err != nil {
-		return false, err
-	}
-
-	if putData {
-		var putFiles int64 = filesPerBlockMax
-		err = j.accumulateBlock(int64(len(buf)), putFiles)
-		if err != nil {
-			return false, err
-		}
+		return err
 	}
 
 	_, err = j.appendJournalEntry(ctx, blockJournalEntry{
 		Op:       blockPutOp,
 		Contexts: kbfsblock.ContextMap{id: {context}},
 	})
-	if err != nil {
-		return false, err
-	}
-
-	return putData, nil
+	return err
 }
 
 func (j *blockJournal) addReference(
@@ -469,7 +484,7 @@ func (j *blockJournal) addReference(
 		return err
 	}
 
-	err = j.s.addReference(id, context, next.String())
+	err = j.s.addReference(ctx, id, context, next.String())
 	if err != nil {
 		return err
 	}
@@ -500,7 +515,7 @@ func (j *blockJournal) archiveReferences(
 		return err
 	}
 
-	err = j.s.archiveReferences(contexts, next.String())
+	err = j.s.archiveReferences(ctx, contexts, next.String())
 	if err != nil {
 		return err
 	}
@@ -545,7 +560,7 @@ func (j *blockJournal) removeReferences(
 		// Remove the references unconditionally here (i.e.,
 		// with an empty tag), since j.s should reflect the
 		// most recent state.
-		liveCount, err := j.s.removeReferences(id, idContexts, "")
+		liveCount, err := j.s.removeReferences(ctx, id, idContexts, "")
 		if err != nil {
 			return nil, err
 		}
@@ -693,7 +708,7 @@ func (j *blockJournal) getNextEntriesToFlush(
 					kbfsmd.RevisionUninitialized, err
 			}
 
-			data, serverHalf, err = j.s.getData(id)
+			data, serverHalf, err = j.s.getData(ctx, id)
 			if err != nil {
 				return blockEntriesToFlush{}, 0,
 					kbfsmd.RevisionUninitialized, err
@@ -851,12 +866,12 @@ func (j *blockJournal) removeFlushedEntry(ctx context.Context,
 			return 0, err
 		}
 
-		err = j.s.markFlushed(id)
+		err = j.s.markFlushed(ctx, id)
 		if err != nil {
 			return 0, err
 		}
 
-		flushedBytes, err = j.s.getDataSize(id)
+		flushedBytes, err = j.s.getDataSize(ctx, id)
 		if err != nil {
 			return 0, err
 		}
@@ -874,7 +889,7 @@ func (j *blockJournal) removeFlushedEntry(ctx context.Context,
 	// references).
 	for id, idContexts := range entry.Contexts {
 		liveCount, err := j.s.removeReferences(
-			id, idContexts, earliestOrdinal.String())
+			ctx, id, idContexts, earliestOrdinal.String())
 		if err != nil {
 			return 0, err
 		}
@@ -974,7 +989,7 @@ func (j *blockJournal) ignoreBlocksAndMDRevMarkersInJournal(ctx context.Context,
 			if e.Op == blockPutOp && isMainJournal {
 				// Treat ignored put ops as flushed
 				// for the purposes of accounting.
-				ignoredBytes, err := j.s.getDataSize(id)
+				ignoredBytes, err := j.s.getDataSize(ctx, id)
 				if err != nil {
 					return 0, err
 				}
@@ -1090,7 +1105,7 @@ func (j *blockJournal) doGC(ctx context.Context,
 		for id := range entry.Contexts {
 			// TODO: once we support references, this needs to be made
 			// goroutine-safe.
-			hasRef, err := j.s.hasAnyRef(id)
+			hasRef, err := j.s.hasAnyRef(ctx, id)
 			if err != nil {
 				return 0, 0, err
 			}