--- a/go/kbfs/libkbfs/block_ops.go
+++ b/go/kbfs/libkbfs/block_ops.go
@@ -65,7 +65,7 @@ func (b *BlockOpsStandard) Get(ctx context.Context, kmd libkey.KeyMetadata,
 	// the block-fetching queue.
 	if journalBServer, ok := b.config.BlockServer().(journalBlockServer); ok {
 		data, serverHalf, found, err := journalBServer.getBlockFromJournal(
-			kmd.TlfID(), blockPtr.ID)
+			ctx, kmd.TlfID(), blockPtr.ID)
 		if err != nil {
 			return err
 		}
@@ -96,7 +96,7 @@ func (b *BlockOpsStandard) GetEncodedSize(ctx context.Context, kmd libkey.KeyMet
 	// the block-fetching queue.
 	if journalBServer, ok := b.config.BlockServer().(journalBlockServer); ok {
 		size, found, err := journalBServer.getBlockSizeFromJournal(
-			kmd.TlfID(), blockPtr.ID)
+			ctx, kmd.TlfID(), blockPtr.ID)
 		if err != nil {
 			return 0, 0, err
 		}