This allows the `SyncAll` code to write the data for multiple blocks in parallel to the disk (but still appends  knowledge of their existence to the journal serially).

Since `blockDiskStore` relied on the upper layer taking a lock during block puts, this required adding rudimentary, per-id exclusivity guarantees from the blockDiskStore, which required plumbing contexts everywhere.

Also, I added a new benchmark that uses big files (instead of lots of small ones) to show the promise of this change.  In the new benchmark, this change resulted in a 20% performance boost for the file-writing phase, and a 15% boost overall when flushing is done concurrently.  Here are results from my new desktop, which has fast M2 SSD memory:

Before this change:

```sh
$ go test -c -tags fuse && ./test.test -test.v -test.bench "BenchmarkMultiFileSyncBigFiles/MDVer\(ImplicitTeams\)" -test.run ^$ -test.benchtime 5s
goos: linux
goarch: amd64
pkg: github.com/keybase/client/go/kbfs/test
BenchmarkMultiFileSyncBigFilesWrites/MDVer(ImplicitTeams)-8         	      50	 126566790 ns/op
BenchmarkMultiFileSyncBigFilesFlush/MDVer(ImplicitTeams)-8          	     100	  62642402 ns/op
BenchmarkMultiFileSyncBigFiles/MDVer(ImplicitTeams)-8               	      50	 169649031 ns/op
PASS
```

After this change:

```sh
$ go test -c -tags fuse && ./test.test -test.v -test.bench "BenchmarkMultiFileSyncBigFiles/MDVer\(ImplicitTeams\)" -test.run ^$ -test.benchtime 5s
goos: linux
goarch: amd64
pkg: github.com/keybase/client/go/kbfs/test
BenchmarkMultiFileSyncBigFilesWrites/MDVer(ImplicitTeams)-8         	     100	 100686847 ns/op
BenchmarkMultiFileSyncBigFilesFlush/MDVer(ImplicitTeams)-8          	     100	  63893657 ns/op
BenchmarkMultiFileSyncBigFiles/MDVer(ImplicitTeams)-8               	      50	 145764578 ns/op
PASS
```

Further, I tested a real-world benchmark of copying the 501 MB from `/keybase/team/keybase/pictures` into KBFS, in the default configuration where flushing happens concurrently.  I saw a nearly 30% boost in that case:

Before this change:
```sh
$ time rsync -atrvp pictures /keybase/private/strib/tmp/pictures-test-1
...
sent 525,053,620 bytes  received 2,498 bytes  36,210,766.76 bytes/sec
total size is 524,916,752  speedup is 1.00

real	0m14.130s
user	0m1.877s
sys	0m0.541s
```

After this change:
```sh
$ time rsync -atrvp pictures /keybase/private/strib/tmp/pictures-test-2
...
sent 525,053,620 bytes  received 2,498 bytes  50,005,344.57 bytes/sec
total size is 524,916,752  speedup is 1.00

real	0m10.063s
user	0m1.728s
sys	0m0.566s
```


Issue: KBFS-3980